Extract ICMP echo responders from lwAFTR
