+++ src/program/snabbvmx/lwaftr/setup.lua
@@ -7,10 +7,10 @@ local basic_apps = require("apps.basic.basic_apps")
 local bt = require("apps.lwaftr.binding_table")
 local config = require("core.config")
 local ethernet = require("lib.protocol.ethernet")
-local ipv4_apps = require("apps.lwaftr.ipv4_apps")
+local ipv4_echo = require("apps.ipv4.echo")
 local ipv4_fragment = require("apps.ipv4.fragment")
 local ipv4_reassemble = require("apps.ipv4.reassemble")
-local ipv6_apps = require("apps.lwaftr.ipv6_apps")
+local ipv6_echo = require("apps.ipv6.echo")
 local ipv6_fragment = require("apps.ipv6.fragment")
 local ipv6_reassemble = require("apps.ipv6.reassemble")
 local lib = require("core.lib")
